Refactor character stage state
