+++ src/pages/IndexPage/IndexPage.tsx
@@ -51,13 +51,9 @@ export function IndexPage() {
             />
           </figure>
           <div className="is-flex is-flex-direction-column is-align-items-center action-buttons">
-            <button
-              className="button is-flex-grow-0 mt-5"
-              title="Coming soon..."
-              disabled
-            >
+            <Link to="/create" className="button is-flex-grow-0 mt-5">
               Create
-            </button>
+            </Link>
             <Link to="/saved" className="button is-flex-grow-0 mt-5">
               Saved
             </Link>
